pyproject.toml
[project]
name = "llama-stack-provider-trustyai-garak"
version = "0.1.8"
description = "Out-Of-Tree Llama Stack provider for Garak Red-teaming"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    {name = "Sai Chandra Pandraju", email = "saichandrapandraju@gmail.com"},
    {name = "TrustyAI Team"}
]
keywords = ["llama-stack", "garak", "red-teaming", "security", "ai-safety"]
dependencies = [
    "llama-stack[client]>=0.4.0",
    # Remote execution dependencies
    "kfp>=2.14.6",
    "kfp-kubernetes>=2.14.6",
    "kfp-server-api>=2.14.6",
    "boto3>=1.35.88",
]

[project.urls]
homepage = "https://github.com/trustyai-explainability/llama-stack-provider-trustyai-garak"
repository = "https://github.com/trustyai-explainability/llama-stack-provider-trustyai-garak"

[build-system]
requires = ["setuptools==80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
inline = [
    "langchain==0.3.27",  # latest langchain breaks garak, so must be '>=0.3.25,<1.0.0'
    "garak==0.12.0",
]
dev = ["pytest", "pytest-cov",  "pytest-asyncio", "black", "isort"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_paths = ["src"]
addopts = "-v"
